feat(ledger): Moved the test to state_test and modified the added db funciton comments
